move menu model location due to circular imports
