+++ napari/_qt/menus/_util.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING, Callable, List, Union
 
 from qtpy.QtWidgets import QAction, QMenu
 
-from ...components.menu import ActionMenuItem, CheckableMenuItem, Menu
+from ...utils.menus import ActionMenuItem, CheckableMenuItem, Menu
 
 if TYPE_CHECKING:
     from typing_extensions import TypedDict
